// src/app/shared/models/userInfo.ts
export interface userInfo {
    user_id: string;
    user_username: string;
    user_email: string;
    user_first_name: string;
    user_last_name: string;
    user_gender: string;
    user_birthday: string;
    user_phone_number: string;
    user_role: string;
    user_address: string;
    user_postcode: string;
    user_city: string;
    user_state: string;
}